Relocate kubernetes secret code
