+++ internal/providers/azure/pke/driver/cluster_deleter.go
@@ -34,7 +34,7 @@ import (
 	"github.com/banzaicloud/pipeline/internal/global"
 	"github.com/banzaicloud/pipeline/internal/providers/azure/pke"
 	"github.com/banzaicloud/pipeline/internal/providers/azure/pke/workflow"
-	intSecret "github.com/banzaicloud/pipeline/internal/secret"
+	"github.com/banzaicloud/pipeline/internal/secret/kubesecret"
 	pkgCluster "github.com/banzaicloud/pipeline/pkg/cluster"
 	"github.com/banzaicloud/pipeline/pkg/k8sclient"
 	pkgAzure "github.com/banzaicloud/pipeline/pkg/providers/azure"
@@ -337,7 +337,7 @@ func gatherK8sServicePublicIPs(publicAddresses []network.PublicIPAddress, cluste
 		return names, nil
 	}
 
-	k8sConfig, err := intSecret.MakeKubeSecretStore(secrets).Get(cluster.OrganizationID, cluster.K8sSecretID)
+	k8sConfig, err := kubesecret.MakeKubeSecretStore(secrets).Get(cluster.OrganizationID, cluster.K8sSecretID)
 	if err != nil {
 		return names, errors.WrapIf(err, "failed to get k8s config")
 	}
